Bring try_extract_score_threshold into collect_filters
