+++ pg_search/src/scan/execution_plan.rs
@@ -35,12 +35,9 @@ use arrow_schema::{SchemaRef, SortOptions};
 use datafusion::common::stats::{ColumnStatistics, Precision};
 use datafusion::common::{DataFusionError, Result, Statistics};
 use datafusion::execution::{RecordBatchStream, SendableRecordBatchStream, TaskContext};
-use datafusion::logical_expr::Operator;
 use datafusion::physical_expr::{EquivalenceProperties, PhysicalExpr, PhysicalSortExpr};
 use datafusion::physical_plan::execution_plan::{Boundedness, EmissionType};
-use datafusion::physical_plan::expressions::{
-    BinaryExpr, Column, DynamicFilterPhysicalExpr, IsNullExpr, Literal,
-};
+use datafusion::physical_plan::expressions::{Column, DynamicFilterPhysicalExpr};
 use datafusion::physical_plan::filter_pushdown::{
     ChildPushdownResult, FilterPushdownPhase, FilterPushdownPropagation, PushedDown,
 };
@@ -50,7 +47,6 @@ use datafusion::physical_plan::metrics::{
 use datafusion::physical_plan::{
     DisplayAs, DisplayFormatType, ExecutionPlan, Partitioning, PlanProperties,
 };
-use datafusion::scalar::ScalarValue;
 use futures::Stream;
 use tantivy::index::SegmentId;
 use tantivy::Score;
@@ -856,120 +852,25 @@ fn build_filters(
     for df in dynamic_filters {
         if let Some(dynamic) = df.downcast_ref::<DynamicFilterPhysicalExpr>() {
             if let Ok(current_expr) = dynamic.current() {
-                if let Some(threshold) =
-                    try_extract_score_threshold(&current_expr, score_col_schema_idx)
-                {
-                    debug_assert!(
-                        score_threshold.is_none(),
-                        "Multiple score thresholds found where we only expect one."
-                    );
-                    score_threshold = Some(threshold);
-                }
-                collect_filters(&current_expr, schema, &mut filters);
+                collect_filters(
+                    &current_expr,
+                    schema,
+                    &mut filters,
+                    score_col_schema_idx,
+                    &mut score_threshold,
+                );
             }
         } else {
-            collect_filters(df, schema, &mut filters);
+            collect_filters(
+                df,
+                schema,
+                &mut filters,
+                score_col_schema_idx,
+                &mut score_threshold,
+            );
         }
     }
     (filters, score_threshold)
-}
-
-/// Check for expressions that we know always evaluate to false
-fn expr_always_false(expr: &Arc<dyn PhysicalExpr>, score_col_schema_idx: usize) -> bool {
-    // FALSE literals are obviously always false
-    if let Some(lit) = expr.downcast_ref::<Literal>() {
-        return matches!(lit.value(), ScalarValue::Boolean(Some(false)));
-    }
-    // score is never null, so 'score IS NULL' is always false
-    if let Some(is_null_expr) = expr.downcast_ref::<IsNullExpr>() {
-        if let Some(col) = is_null_expr.arg().downcast_ref::<Column>() {
-            return col.index() == score_col_schema_idx;
-        }
-    }
-    false
-}
-
-/// Attempt to extract a minimum score threshold from the expression. Bounds propagate as:
-/// - `score > t`  => `t`
-/// - `score = t`  => `t.next_down()` (rows at exactly `t` must survive)
-/// - `AND`        => the minimum of the bounds from either side (so it holds for the whole conjunction)
-/// - `OR`  
-///     - If both sides contain a bound, keep the minimum. If one side has a bound, use
-///       it only in the case the other side always evaluates to false. Any OR with a
-///       possibly-true non-score expression cannot provide a threshold, as the non-score
-///       arm may be true
-///
-/// The returned threshold value assumes the threshold check uses > (greater-than) semantics.
-/// ASSUMPTION: We intentionally don't check the Operator::Lt variant as DataFusion always puts the column
-/// on the left.
-///
-/// This is necessary for using the blockmax-wand optimization in joins
-fn try_extract_score_threshold(
-    expr: &Arc<dyn PhysicalExpr>,
-    score_col_schema_idx: Option<usize>,
-) -> Option<Score> {
-    let score_col_schema_idx = score_col_schema_idx?;
-    let binary_expr = expr.downcast_ref::<BinaryExpr>()?;
-    match binary_expr.op() {
-        Operator::And => match (
-            try_extract_score_threshold(binary_expr.left(), Some(score_col_schema_idx)),
-            try_extract_score_threshold(binary_expr.right(), Some(score_col_schema_idx)),
-        ) {
-            // We should never see a score threshold on both sides of an AND, but
-            // if we do, this will at least return a value that is safe to prune.
-            // We might have threshold's on both sides of an OR. In both cases,
-            // we can't take the higher value, as the lower value may come from an
-            // Eq check
-            (Some(left), Some(right)) => Some(left.min(right)),
-            (Some(left), None) => Some(left),
-            (None, Some(right)) => Some(right),
-            (None, None) => None,
-        },
-        Operator::Gt => {
-            // Look for a binary expr that looks like: score > f32
-            let col = binary_expr.left().downcast_ref::<Column>()?;
-            if col.index() != score_col_schema_idx {
-                return None;
-            }
-            match binary_expr.right().downcast_ref::<Literal>()?.value() {
-                ScalarValue::Float32(Some(t)) => Some(*t),
-                _ => None,
-            }
-        }
-        Operator::Eq => {
-            // Look for a binary expr that looks like: score = f32
-            let col = binary_expr.left().downcast_ref::<Column>()?;
-            if col.index() != score_col_schema_idx {
-                return None;
-            }
-            match binary_expr.right().downcast_ref::<Literal>()?.value() {
-                // PruningScorer's assume the threshold has greater-than semantics, so
-                // take the next representable value below the eq check to keep the threshold valid
-                ScalarValue::Float32(Some(t)) => Some(t.next_down()),
-                _ => None,
-            }
-        }
-        Operator::Or => {
-            // Members of an OR expression that are always false can be safely ignored, so we can
-            // try to pull the score threshold from the other side.
-            if expr_always_false(binary_expr.left(), score_col_schema_idx) {
-                try_extract_score_threshold(binary_expr.right(), Some(score_col_schema_idx))
-            } else if expr_always_false(binary_expr.right(), score_col_schema_idx) {
-                try_extract_score_threshold(binary_expr.left(), Some(score_col_schema_idx))
-            }
-            // If both sides have a threshold-containing part, take the lower threshold: then any matching
-            // row satisfies one of the arms and therefore exceeds the smaller bound, so pruning by it is safe
-            else if let (Some(left), Some(right)) = (
-                try_extract_score_threshold(binary_expr.left(), Some(score_col_schema_idx)),
-                try_extract_score_threshold(binary_expr.right(), Some(score_col_schema_idx)),
-            ) {
-                Some(left.min(right))
-            } else {
-                None
-            }
-        }
-        _ => None,
-    }
 }
 
 /// A wrapper that unsafely implements Send for a Stream.
@@ -1010,146 +911,14 @@ mod tests {
     use std::sync::Arc;
 
     use arrow_schema::{Schema, SchemaRef};
-    use datafusion::logical_expr::Operator;
-    use datafusion::physical_expr::expressions::{is_not_null, is_null, lit, BinaryExpr, Column};
-    use datafusion::physical_expr::PhysicalExpr;
-    use datafusion::scalar::ScalarValue;
     use pgrx::prelude::*;
 
     use crate::query::SearchQueryInput;
 
-    use super::{try_extract_score_threshold, PgSearchScanPlan};
+    use super::PgSearchScanPlan;
 
     fn empty_schema() -> SchemaRef {
         Arc::new(Schema::empty())
-    }
-
-    const SCORE_IDX: usize = 0;
-    const ID_IDX: usize = 1;
-
-    fn score() -> Arc<dyn PhysicalExpr> {
-        Arc::new(Column::new("pdb.score()", SCORE_IDX))
-    }
-
-    fn id() -> Arc<dyn PhysicalExpr> {
-        Arc::new(Column::new("id", ID_IDX))
-    }
-
-    fn f32_lit(v: f32) -> Arc<dyn PhysicalExpr> {
-        lit(ScalarValue::Float32(Some(v)))
-    }
-
-    fn int_lit(v: i64) -> Arc<dyn PhysicalExpr> {
-        lit(ScalarValue::Int64(Some(v)))
-    }
-
-    fn bin(
-        left: Arc<dyn PhysicalExpr>,
-        op: Operator,
-        right: Arc<dyn PhysicalExpr>,
-    ) -> Arc<dyn PhysicalExpr> {
-        Arc::new(BinaryExpr::new(left, op, right))
-    }
-
-    #[test]
-    fn score_threshold_bare_gt_is_exact() {
-        // Single-key `ORDER BY score DESC` publish. The extracted threshold is the
-        // unrelaxed t: with no tiebreakers, a row tied with the cutoff can never
-        // displace it, so the scorer's strict `score > t` mirrors the filter exactly.
-        let expr = bin(score(), Operator::Gt, f32_lit(1.5));
-        assert_eq!(
-            try_extract_score_threshold(&expr, Some(SCORE_IDX)),
-            Some(1.5f32)
-        );
-    }
-
-    #[test]
-    fn score_threshold_lexicographic_chain_with_score_leading() {
-        // `ORDER BY score DESC, id ASC` publish: score > t OR (score = t AND id < v).
-        // Rows tied at t may still win on the tiebreaker, so the extracted bound is
-        // next_down(t): under the scorer's strict `>` semantics that keeps score >= t.
-        let chain = bin(
-            bin(score(), Operator::Gt, f32_lit(1.5)),
-            Operator::Or,
-            bin(
-                bin(score(), Operator::Eq, f32_lit(1.5)),
-                Operator::And,
-                bin(id(), Operator::Lt, int_lit(10)),
-            ),
-        );
-        assert_eq!(
-            try_extract_score_threshold(&chain, Some(SCORE_IDX)),
-            Some(1.5f32.next_down())
-        );
-    }
-
-    #[test]
-    fn score_threshold_nulls_first_wrapper() {
-        // DESC NULLS FIRST publish: score IS NULL OR score > t. The IS NULL arm can
-        // never match (every scored doc has a score), so the bound still holds.
-        let expr = bin(
-            is_null(score()).unwrap(),
-            Operator::Or,
-            bin(score(), Operator::Gt, f32_lit(1.5)),
-        );
-        assert_eq!(
-            try_extract_score_threshold(&expr, Some(SCORE_IDX)),
-            Some(1.5f32)
-        );
-    }
-
-    #[test]
-    fn score_threshold_conjunctive_null_guard() {
-        // DESC NULLS LAST publish: score IS NOT NULL AND score > t. A conjunct's
-        // bound holds for the whole conjunction, so extraction is sound.
-        let expr = bin(
-            is_not_null(score()).unwrap(),
-            Operator::And,
-            bin(score(), Operator::Gt, f32_lit(1.5)),
-        );
-        assert_eq!(
-            try_extract_score_threshold(&expr, Some(SCORE_IDX)),
-            Some(1.5f32)
-        );
-    }
-
-    #[test]
-    fn score_threshold_rejected_when_score_is_tiebreaker() {
-        // `ORDER BY id ASC, score DESC` publish: id < v OR (id = v AND score > t).
-        // The left arm admits rows of ANY score, so the expression implies no score
-        // bound; extracting t here would block-prune rows that match via `id < v`.
-        let chain = bin(
-            bin(id(), Operator::Lt, int_lit(10)),
-            Operator::Or,
-            bin(
-                bin(id(), Operator::Eq, int_lit(10)),
-                Operator::And,
-                bin(score(), Operator::Gt, f32_lit(1.5)),
-            ),
-        );
-        assert_eq!(try_extract_score_threshold(&chain, Some(SCORE_IDX)), None);
-    }
-
-    #[test]
-    fn score_threshold_rejects_foreign_shapes() {
-        // Gt on a non-score column.
-        let other_col = bin(id(), Operator::Gt, f32_lit(1.5));
-        assert_eq!(
-            try_extract_score_threshold(&other_col, Some(SCORE_IDX)),
-            None
-        );
-
-        // Non-Float32 literal: the score column is always Float32, so a Float64
-        // comparison (e.g. cast-normalized by an optimizer pass) must not parse.
-        let f64_cmp = bin(score(), Operator::Gt, lit(ScalarValue::Float64(Some(1.5))));
-        assert_eq!(try_extract_score_threshold(&f64_cmp, Some(SCORE_IDX)), None);
-
-        // The lit(true) placeholder every dynamic filter holds before the first
-        // TopK publish.
-        assert_eq!(
-            try_extract_score_threshold(&lit(true), Some(SCORE_IDX)),
-            None
-        );
     }
 
     #[pg_test]
