+++ pg_search/src/scan/pre_filter.rs
@@ -101,7 +101,7 @@ use datafusion::physical_expr::expressions::{
 use datafusion::physical_expr::PhysicalExpr;
 use datafusion::physical_plan::expressions::InListExpr;
 use datafusion::physical_plan::joins::HashTableLookupExpr;
-use tantivy::SegmentOrdinal;
+use tantivy::{Score, SegmentOrdinal};
 
 use crate::api::HashSet;
 use crate::index::fast_fields_helper::{FFHelper, FFType, NULL_TERM_ORDINAL};
@@ -261,15 +261,44 @@ impl PreFilter {
 /// Top-level `AND` operations are split into separate `PreFilter`s to allow early
 /// short-circuiting in the scanner. Expressions containing unsupported nodes
 /// (e.g. non-comparison operators, functions) are safely skipped.
-pub fn collect_filters(expr: &Arc<dyn PhysicalExpr>, schema: &SchemaRef, out: &mut Vec<PreFilter>) {
+pub fn collect_filters(
+    expr: &Arc<dyn PhysicalExpr>,
+    schema: &SchemaRef,
+    out: &mut Vec<PreFilter>,
+    score_col_schema_idx: Option<usize>,
+    score_threshold: &mut Option<Score>,
+) {
     // Split top-level ANDs to maximize early pruning
     if let Some(binary) = expr.downcast_ref::<BinaryExpr>() {
         if matches!(binary.op(), Operator::And) {
-            collect_filters(binary.left(), schema, out);
-            collect_filters(binary.right(), schema, out);
+            collect_filters(
+                binary.left(),
+                schema,
+                out,
+                score_col_schema_idx,
+                score_threshold,
+            );
+            collect_filters(
+                binary.right(),
+                schema,
+                out,
+                score_col_schema_idx,
+                score_threshold,
+            );
             return;
         }
     }
+
+    let threshold = match (
+        try_extract_score_threshold(expr, score_col_schema_idx),
+        &score_threshold,
+    ) {
+        (Some(new), Some(existing)) => Some(new.min(*existing)),
+        (Some(new), None) => Some(new),
+        (None, Some(existing)) => Some(*existing),
+        (None, None) => None,
+    };
+    *score_threshold = threshold;
 
     // Check if the expression is supported for pre-filtering
     let mut required_columns = Vec::new();
@@ -280,6 +309,100 @@ pub fn collect_filters(expr: &Arc<dyn PhysicalExpr>, schema: &SchemaRef, out: &m
             expr: Arc::clone(expr),
             required_columns,
         });
+    }
+}
+
+/// Check for expressions that we know always evaluate to false
+fn expr_always_false(expr: &Arc<dyn PhysicalExpr>, score_col_schema_idx: usize) -> bool {
+    // FALSE literals are obviously always false
+    if let Some(lit) = expr.downcast_ref::<Literal>() {
+        return matches!(lit.value(), ScalarValue::Boolean(Some(false)));
+    }
+    // score is never null, so 'score IS NULL' is always false
+    if let Some(is_null_expr) = expr.downcast_ref::<IsNullExpr>() {
+        if let Some(col) = is_null_expr.arg().downcast_ref::<Column>() {
+            return col.index() == score_col_schema_idx;
+        }
+    }
+    false
+}
+
+/// Attempt to extract a minimum score threshold from the expression. Bounds propagate as:
+/// - `score > t`  => `t`
+/// - `score = t`  => `t.next_down()` (rows at exactly `t` must survive)
+/// - `AND`        => the minimum of the bounds from either side (so it holds for the whole conjunction)
+/// - `OR`  
+///     - If both sides contain a bound, keep the minimum. If one side has a bound, use
+///       it only in the case the other side always evaluates to false. Any OR with a
+///       possibly-true non-score expression cannot provide a threshold, as the non-score
+///       arm may be true
+///
+/// The returned threshold value assumes the threshold check uses > (greater-than) semantics.
+/// ASSUMPTION: We intentionally don't check the Operator::Lt variant as DataFusion always puts the column
+/// on the left.
+///
+/// This is necessary for using the blockmax-wand optimization in joins
+fn try_extract_score_threshold(
+    expr: &Arc<dyn PhysicalExpr>,
+    score_col_schema_idx: Option<usize>,
+) -> Option<Score> {
+    let score_col_schema_idx = score_col_schema_idx?;
+    let binary_expr = expr.downcast_ref::<BinaryExpr>()?;
+    match binary_expr.op() {
+        Operator::Gt => {
+            // Look for a binary expr that looks like: score > f32
+            let col = binary_expr.left().downcast_ref::<Column>()?;
+            if col.index() != score_col_schema_idx {
+                return None;
+            }
+            match binary_expr.right().downcast_ref::<Literal>()?.value() {
+                ScalarValue::Float32(Some(t)) => Some(*t),
+                _ => None,
+            }
+        }
+        Operator::Eq => {
+            // Look for a binary expr that looks like: score = f32
+            let col = binary_expr.left().downcast_ref::<Column>()?;
+            if col.index() != score_col_schema_idx {
+                return None;
+            }
+            match binary_expr.right().downcast_ref::<Literal>()?.value() {
+                // PruningScorer's assume the threshold has greater-than semantics, so
+                // take the next representable value below the eq check to keep the threshold valid
+                ScalarValue::Float32(Some(t)) => Some(t.next_down()),
+                _ => None,
+            }
+        }
+        Operator::And => {
+            match (
+                try_extract_score_threshold(binary_expr.left(), Some(score_col_schema_idx)),
+                try_extract_score_threshold(binary_expr.right(), Some(score_col_schema_idx)),
+            ) {
+                (Some(left), Some(right)) => Some(left.min(right)),
+                (Some(v), None) | (None, Some(v)) => Some(v),
+                (None, None) => None,
+            }
+        }
+        Operator::Or => {
+            // Members of an OR expression that are always false can be safely ignored, so we can
+            // try to pull the score threshold from the other side.
+            if expr_always_false(binary_expr.left(), score_col_schema_idx) {
+                try_extract_score_threshold(binary_expr.right(), Some(score_col_schema_idx))
+            } else if expr_always_false(binary_expr.right(), score_col_schema_idx) {
+                try_extract_score_threshold(binary_expr.left(), Some(score_col_schema_idx))
+            }
+            // If both sides have a threshold-containing part, take the lower threshold: then any matching
+            // row satisfies one of the arms and therefore exceeds the smaller bound, so pruning by it is safe
+            else if let (Some(left), Some(right)) = (
+                try_extract_score_threshold(binary_expr.left(), Some(score_col_schema_idx)),
+                try_extract_score_threshold(binary_expr.right(), Some(score_col_schema_idx)),
+            ) {
+                Some(left.min(right))
+            } else {
+                None
+            }
+        }
+        _ => None,
     }
 }
 
@@ -835,5 +958,143 @@ pub fn try_dynamic_filter_pushdown(
         let boolean_query = BooleanQuery::new(combined_musts);
         reader.and_query(Box::new(boolean_query));
         true
+    }
+}
+
+#[cfg(any(test, feature = "pg_test"))]
+mod tests {
+    use super::try_extract_score_threshold;
+    use datafusion::logical_expr::Operator;
+    use datafusion::physical_expr::expressions::{is_not_null, is_null, lit, BinaryExpr, Column};
+    use datafusion::physical_expr::PhysicalExpr;
+    use datafusion::scalar::ScalarValue;
+    use std::sync::Arc;
+
+    const SCORE_IDX: usize = 0;
+    const ID_IDX: usize = 1;
+
+    fn score() -> Arc<dyn PhysicalExpr> {
+        Arc::new(Column::new("pdb.score()", SCORE_IDX))
+    }
+
+    fn id() -> Arc<dyn PhysicalExpr> {
+        Arc::new(Column::new("id", ID_IDX))
+    }
+
+    fn f32_lit(v: f32) -> Arc<dyn PhysicalExpr> {
+        lit(ScalarValue::Float32(Some(v)))
+    }
+
+    fn int_lit(v: i64) -> Arc<dyn PhysicalExpr> {
+        lit(ScalarValue::Int64(Some(v)))
+    }
+
+    fn bin(
+        left: Arc<dyn PhysicalExpr>,
+        op: Operator,
+        right: Arc<dyn PhysicalExpr>,
+    ) -> Arc<dyn PhysicalExpr> {
+        Arc::new(BinaryExpr::new(left, op, right))
+    }
+
+    #[test]
+    fn score_threshold_bare_gt_is_exact() {
+        // Single-key `ORDER BY score DESC` publish. The extracted threshold is the
+        // unrelaxed t: with no tiebreakers, a row tied with the cutoff can never
+        // displace it, so the scorer's strict `score > t` mirrors the filter exactly.
+        let expr = bin(score(), Operator::Gt, f32_lit(1.5));
+        assert_eq!(
+            try_extract_score_threshold(&expr, Some(SCORE_IDX)),
+            Some(1.5f32)
+        );
+    }
+
+    #[test]
+    fn score_threshold_lexicographic_chain_with_score_leading() {
+        // `ORDER BY score DESC, id ASC` publish: score > t OR (score = t AND id < v).
+        // Rows tied at t may still win on the tiebreaker, so the extracted bound is
+        // next_down(t): under the scorer's strict `>` semantics that keeps score >= t.
+        let chain = bin(
+            bin(score(), Operator::Gt, f32_lit(1.5)),
+            Operator::Or,
+            bin(
+                bin(score(), Operator::Eq, f32_lit(1.5)),
+                Operator::And,
+                bin(id(), Operator::Lt, int_lit(10)),
+            ),
+        );
+        assert_eq!(
+            try_extract_score_threshold(&chain, Some(SCORE_IDX)),
+            Some(1.5f32.next_down())
+        );
+    }
+
+    #[test]
+    fn score_threshold_nulls_first_wrapper() {
+        // DESC NULLS FIRST publish: score IS NULL OR score > t. The IS NULL arm can
+        // never match (every scored doc has a score), so the bound still holds.
+        let expr = bin(
+            is_null(score()).unwrap(),
+            Operator::Or,
+            bin(score(), Operator::Gt, f32_lit(1.5)),
+        );
+        assert_eq!(
+            try_extract_score_threshold(&expr, Some(SCORE_IDX)),
+            Some(1.5f32)
+        );
+    }
+
+    #[test]
+    fn score_threshold_conjunctive_null_guard() {
+        // DESC NULLS LAST publish: score IS NOT NULL AND score > t. A conjunct's
+        // bound holds for the whole conjunction, so extraction is sound.
+        let expr = bin(
+            is_not_null(score()).unwrap(),
+            Operator::And,
+            bin(score(), Operator::Gt, f32_lit(1.5)),
+        );
+        assert_eq!(
+            try_extract_score_threshold(&expr, Some(SCORE_IDX)),
+            Some(1.5f32)
+        );
+    }
+
+    #[test]
+    fn score_threshold_rejected_when_score_is_tiebreaker() {
+        // `ORDER BY id ASC, score DESC` publish: id < v OR (id = v AND score > t).
+        // The left arm admits rows of ANY score, so the expression implies no score
+        // bound; extracting t here would block-prune rows that match via `id < v`.
+        let chain = bin(
+            bin(id(), Operator::Lt, int_lit(10)),
+            Operator::Or,
+            bin(
+                bin(id(), Operator::Eq, int_lit(10)),
+                Operator::And,
+                bin(score(), Operator::Gt, f32_lit(1.5)),
+            ),
+        );
+        assert_eq!(try_extract_score_threshold(&chain, Some(SCORE_IDX)), None);
+    }
+
+    #[test]
+    fn score_threshold_rejects_foreign_shapes() {
+        // Gt on a non-score column.
+        let other_col = bin(id(), Operator::Gt, f32_lit(1.5));
+        assert_eq!(
+            try_extract_score_threshold(&other_col, Some(SCORE_IDX)),
+            None
+        );
+
+        // Non-Float32 literal: the score column is always Float32, so a Float64
+        // comparison (e.g. cast-normalized by an optimizer pass) must not parse.
+        let f64_cmp = bin(score(), Operator::Gt, lit(ScalarValue::Float64(Some(1.5))));
+        assert_eq!(try_extract_score_threshold(&f64_cmp, Some(SCORE_IDX)), None);
+
+        // The lit(true) placeholder every dynamic filter holds before the first
+        // TopK publish.
+        assert_eq!(
+            try_extract_score_threshold(&lit(true), Some(SCORE_IDX)),
+            None
+        );
     }
 }
